style: Rename util files
- 디렉토리 구조 수정으로 인하여 파일 이름 변경
- 해당 모듈을 import하는 파일도 함께 경로 수정
