+++ src/components/common/Dialog/Dialog.tsx
@@ -3,7 +3,7 @@
 import { useCallback, useEffect, useRef } from 'react';
 import { createPortal } from 'react-dom';
 import styles from './Dialog.module.css';
-import { getTabbableChildren } from '@/utils/client/focus-tab';
+import { getTabbableChildren } from '@/utils/client';
 import { ReactComponent as CancelIcon } from '@svgs/close-svgrepo-com.svg';
 
 export interface DialogProps {
